congregate WindowsUpdateJob functionality into single job, related changes to commands and related types
