+++ src/Public/Install-WindowsUpdate.ps1
@@ -3,9 +3,9 @@ function Install-WindowsUpdate {
     [OutputType([KPBSD.PowerShell.WindowsUpdate.UpdateModel], ParameterSetName = 'TitlePassThru')]
     [OutputType([KPBSD.PowerShell.WindowsUpdate.UpdateModel], ParameterSetName = 'IdPassThru')]
     [OutputType([KPBSD.PowerShell.WindowsUpdate.UpdateModel], ParameterSetName = 'UpdatePassThru')]
-    [OutputType([KPBSD.PowerShell.WindowsUpdate.WUInstallJob], ParameterSetName = 'TitleAsJob')]
-    [OutputType([KPBSD.PowerShell.WindowsUpdate.WUInstallJob], ParameterSetName = 'IdAsJob')]
-    [OutputType([KPBSD.PowerShell.WindowsUpdate.WUInstallJob], ParameterSetName = 'UpdateAsJob')]
+    [OutputType([KPBSD.PowerShell.WindowsUpdate.WindowsUpdateJob], ParameterSetName = 'TitleAsJob')]
+    [OutputType([KPBSD.PowerShell.WindowsUpdate.WindowsUpdateJob], ParameterSetName = 'IdAsJob')]
+    [OutputType([KPBSD.PowerShell.WindowsUpdate.WindowsUpdateJob], ParameterSetName = 'UpdateAsJob')]
     [CmdletBinding(SupportsShouldProcess, DefaultParameterSetName = 'TitlePassThru')]
     param(
         [Parameter(Mandatory, Position = 0, ValueFromPipelineByPropertyName, ParameterSetName = 'TitlePassThru')]
@@ -143,7 +143,14 @@ function Install-WindowsUpdate {
 
         Write-Debug "$(Get-Date -Format 'HH:mm:ss.ffff') [Install-WindowsUpdate] <Process> Proceeding with $($AcceptedUpdates.Count) accepted Windows Updates to install."
         if ($ProcessPipelineAsIndividualJobs -and $AcceptedUpdates.Count -gt 0) {
-            $Job = Start-WindowsUpdateInstallJob -WindowsUpdate $AcceptedUpdates -JobName $PSBoundParameters['JobName'] -Command $MyInvocation.Line
+            $StartWindowsUpdateJobParameters = @{
+                'WindowsUpdate' = $AcceptedUpdates
+                'JobName' = $PSBoundParameters['JobName']
+                'Command' = $MyInvocation.Line
+                'WindowsUpdateSession' = Get-WindowsUpdateSession
+                'Install' = $true
+            }
+            $Job = Start-WindowsUpdateJob @StartWindowsUpdateJobParameters
             if ($AsJob) {
                 $Job
             }
@@ -155,7 +162,14 @@ function Install-WindowsUpdate {
     end {
         Write-Debug "$(Get-Date -Format 'HH:mm:ss.ffff') [Install-WindowsUpdate] <End> Proceeding with $($AcceptedUpdates.Count) accepted Windows Updates to install."
         if (!$ProcessPipelineAsIndividualJobs -and $AcceptedUpdates.Count -gt 0) {
-            $Job = Start-WindowsUpdateInstallJob -WindowsUpdate $AcceptedUpdates -JobName $PSBoundParameters['JobName'] -Command $MyInvocation.Line
+            $StartWindowsUpdateJobParameters = @{
+                'WindowsUpdate' = $AcceptedUpdates
+                'JobName' = $PSBoundParameters['JobName']
+                'Command' = $MyInvocation.Line
+                'WindowsUpdateSession' = Get-WindowsUpdateSession
+                'Install' = $true
+            }
+            $Job = Start-WindowsUpdateJob @StartWindowsUpdateJobParameters
             if ($AsJob) {
                 $Job
             }
